Fixing velocities received from the software

+++ firmware_msc/Electronics_v2-master/Firmware/SSL_Base/Inc/ssl_main.h
@@ -64,7 +64,7 @@ typedef struct _RobotSensors
 	{
 		uint8_t updated;
 		uint32_t time;
-		int8_t theoVel[3];
+		float theoVel[3];
 	} theoreticalVel;
 
 	struct PACKED
